requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
Werkzeug==3.0.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
